be/requirements.txt
fastapi==0.109.0
uvicorn==0.23.2
pandas==2.1.1
numpy==1.26.2
python-dotenv==1.0.0
requests==2.32.0
faster-whisper==0.5.2
google-generativeai==0.1.4
